Use a refinement list for filtering published videos in devenv
This makes it easier to view only unpublished videos or only published
videos, as opposed to using a toggle option. The toggle button had the
advantage of filtering out other filtering options, which I suppose
could still be done with this if we first filtered on the published
attribute, before looking at any of the other ones.