Incrementally crawl data from YouTube
Previously, we used a brute-force approach to crawling the YouTube
channels. This worked well enough, until it suddenly broke - possibly
because YouTube changed things, or because of network differences.

We'd just fetch all the videos for all the channels each time we
updated the data, assuming that this update script would only be run
once a week or so. But, on some networks this approach failed, and
from other IPs it would get rate limited, etc.

This commit moves to an incremental approach to fetching the data,
without completely changing the format in which data is stored,
etc. We do an initial crawl of the channels/playlists to get highlevel
metadata for videos (essentially ids+titles). We then use this
metadata to fetch additional metadata for videos that seem interesting
from their title. We only fetch the additional metadata for "new"
videos -- the videos for which we didn't previously fetch this
metadata. This metadata is stored as flat files in a .json directory
in the data directory.